Rename File.ts to file.ts
